+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Improve commit behavior
 - Move gg commit conventions from gg_git to gg_one_core
+- Answer gg did publish from git tags instead of a marker
 
 ## 2.2.0 - 2026-08-09